requirements.txt
# ─────────────────────────────────────────────────────────────────
# Enterprise RAG Pipeline Dependencies
# ─────────────────────────────────────────────────────────────────

# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP & Async
httpx>=0.27.0
aiofiles>=23.0.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Data Sources
arxiv>=2.1.0
wikipedia-api>=0.6.0
feedparser>=6.0.0

# Language Detection
langdetect>=1.0.9

# Synthetic data generation
faker>=24.0.0

# Utilities
loguru>=0.7.0
rich>=13.0.0
typer>=0.12.0
pyyaml>=6.0.0
tenacity>=8.0.0
orjson>=3.9.0

# AI APIs
openai>=1.0.0
anthropic>=0.25.0

# ── Phase II: Chunking, Embedding, Indexing, Retrieval ───────────────────────

# Tokenization (for chunk-size counting)
tiktoken>=0.7.0

# FAISS vector index (CPU build -- use faiss-gpu if you have CUDA)
faiss-cpu>=1.8.0

# BM25 keyword index (hybrid retrieval)
rank-bm25>=0.2.2

# Numpy (FAISS dependency, explicit pin for Windows compatibility)
numpy>=1.26.0

# Sentence splitting for semantic chunking
nltk>=3.8.0

# LangSmith observability + cost tracking
langsmith>=0.1.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
